Fixed a horribly stupid code raising warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_datatools/trunk@79 f95d2973-9550-45a5-b795-f9f66c6b6f94

+++ src/ParticleDensityOpacityModule.cpp
@@ -615,7 +615,7 @@ void megamol::stdplugin::datatools::ParticleDensityOpacityModule::compute_densit
             float x_move = 0.0f;
             float y_move = 0.0f;
             float z_move = 0.0f;
-            const float range = 1;
+            const int range = 1;
 
             for (int ix = -range; ix <= range; ix++) {
                 int cx = static_cast<int>(x) + ix;
